backend/requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
bleach==6.1.0
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
cryptography==42.0.8
Django==5.0.7
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.2
django-resized==1.0.2
django-rest-knox==4.2.0
django-rest-passwordreset==1.4.1
django-summernote==0.8.20.0
djangorestframework==3.15.2
msgpack==1.0.8
pillow==10.3.0
pycparser==2.22
redis==5.0.7
shortuuid==1.0.13
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
webencodings==0.5.1
